+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- full_name property in Employee Model
 - New tests of the index View
 - Index View of employee Module to show all employees - ***/employee***